07_agents/02_sqlite/requirements.txt
fastmcp
requests
langchain
langgraph
langchain-mcp-adapters
langchain-openai
langchain_google_genai
